fix(app): serialize concurrent speakers.json writes via process-wide lock
Every mutation site (RPC handlers, pipeline-job confirmation, KnownVoices
UI, voice enrollment) constructs a fresh SpeakerMatcher and runs a
load-mutate-save sequence. Without serialization, two writers can both
load the same snapshot and the second saveDB silently overwrites the
first — a lost-update race for speakers.json.

Add a process-wide NSLock + a new mutateDB(_:) helper that wraps
load → block → save under the lock. Refactor renameSpeaker,
deleteSpeaker, mergeSpeakers, updateDB, and the RPC seedSpeaker handler
to go through it. Reads stay unlocked: every code path reloads after a
mutation anyway and replaceItemAt is already atomic at the FS level.

Two regression tests: a direct mutateDB stress test (100 concurrent
appends → final count must equal 101) and a high-level renameSpeaker
stress test (50 concurrent renames of distinct entries → all must
survive). Both deterministically fail on unsynchronized RMW.

Extract StoredSpeaker into its own file — keeps SpeakerMatcher.swift
under the file-length lint budget and matches the codebase pattern of
one type per file.

Refs side-effects review finding H4.
